Put ftrace functions in another helper file and
